add arm64 helm client retrieval support
